cleanup and improvements

+++ lsp/document.go
@@ -122,7 +122,7 @@ func (doc *Document) getAssignedValue(id ast.NodeID) ast.NodeID {
 		return ast.InvalidNode
 	}
 
-	if int(id) < len(doc.Tree.Nodes) {
+	if uint(id) < uint(len(doc.Tree.Nodes)) {
 		kind := doc.Tree.Nodes[id].Kind
 		if kind == ast.KindFunctionExpr || kind == ast.KindTableExpr || kind == ast.KindString || kind == ast.KindNumber || kind == ast.KindTrue || kind == ast.KindFalse || kind == ast.KindNil {
 			return id
